시간 형식이 자바의 interger보다 큼

+++ android/src/main/java/com/ipinlabs/apache_math3/ApacheMath3Plugin.java
@@ -2,6 +2,7 @@ package com.ipinlabs.apache_math3;
 
 import androidx.annotation.NonNull;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import io.flutter.embedding.engine.plugins.FlutterPlugin;
@@ -42,7 +43,7 @@ public class ApacheMath3Plugin implements FlutterPlugin, MethodCallHandler {
     if (call.method.equals("getPlatformVersion")) {
       result.success("Android " + android.os.Build.VERSION.RELEASE);
     }else if(call.method.equals("linearErp")){
-      List<Integer> t_in = call.argument("input");
+      List<BigInteger> t_in = call.argument("input");
 //      List<Double> originValue = call.argument("value");
       List<Integer> originOutput  = call.argument("output");
 
@@ -51,8 +52,8 @@ public class ApacheMath3Plugin implements FlutterPlugin, MethodCallHandler {
       assert  originOutput != null;
 
       List<Long> longTinList = new ArrayList<>();
-      for (Integer inElement : t_in) {
-        longTinList.add(Integer.valueOf(inElement).longValue());
+      for (BigInteger inElement : t_in) {
+        longTinList.add(inElement.longValue());
       }
 
       List<Double> tempValueList = new ArrayList<>();
